This fixes a bug where the optimized sorting routines weren't used on ARM64 (where `purego=true`). 

I've also added benchmarks for indirect sorting, and added tests for pivot selection to make sure we aren't triggering worst case quicksort behavior.