@tysonmote 2022-04-08
Big improvement in `cachette/lsm`'s sorting benchmarks, which I rigged up to sort 100K items that are randomly ordered and ordered ascending and descending:

```
name                            old time/op  new time/op  delta
BufferSort_sortedAscending-10   23.9ms ± 3%   3.4ms ± 2%   ~     (p=0.100 n=3+3)
BufferSort_sortedDescending-10  16.7ms ± 3%   3.6ms ± 2%   ~     (p=0.100 n=3+3)
BufferSort_random-10            9.54ms ± 7%  9.19ms ± 2%   ~     (p=0.700 n=3+3)
```